Problem: Fetching PRs with Github API not working (#190)
Solution: Instead of calling the method from client, convert it to a validated method and call it from a server. Fetch the access token from `Tokens` collection and fallback if there's no token available (endpoints are available without a token, but they're severely rate-limited). Add server-side and client-side tests for the logic.